+++ src/paths.ts
@@ -16,6 +16,7 @@ export const TEMP_DIR = path.join(process.cwd(), TEMP);
 export const CONTRACTS_DIR = path.join(process.cwd(), CONTRACTS);
 export const TESTS_DIR = path.join(process.cwd(), TESTS);
 
+export const PACKAGE_ENTRY_POINT = path.join(process.cwd(), 'package.ts');
 export const BLUEPRINT_CONFIG = path.join(process.cwd(), 'blueprint.config.ts');
 export const TACT_ROOT_CONFIG = path.join(process.cwd(), 'tact.config.json');
 export const TYPESCRIPT_CONFIG = path.join(process.cwd(), 'tsconfig.json');