+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added `tact.config.json` support
+- Added the `pack` command which building and preparing a publish-ready package of contracts wrappers
 
 ## [0.32.1] - 2025-05-06